pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Shazil91 <shazilali312@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.1"}
uvicorn = "^0.31.1"
pydantic = "^2.9.2"
sqlalchemy = "^2.0.36"
psycopg2 = "^2.9.10"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic-settings = "^2.6.0"
bcrypt = "^4.2.0"
pytest = "^8.3.3"
alembic = "^1.14.0"


[tool.poetry.group.dev.dependencies]
types-python-jose = "^3.3.4.20240106"
types-passlib = "^1.7.7.20240819"
types-psycopg2 = "^2.9.21.20241019"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
